refactor: split app.rs into per-screen modules (#19)
Extract the 3,865-line app.rs monolith into focused modules:
- edit.rs (579 lines) — edit overlay draw/key handling + tests
- settings.rs (848 lines) — settings screen draw/key handling + tests
- files.rs (1,554 lines) — files screen draw/key handling + tests
- app.rs (1,023 lines) — App struct, enums, state types, core dispatch

Pure refactor with zero behavior changes. All 176 tests pass.
