+++ src/tui/app.rs
@@ -2,7 +2,6 @@ use std::cell::Cell;
 use std::collections::HashSet;
 use std::fs;
 use std::io;
-use std::io::Write;
 use std::path::Path;
 use std::path::PathBuf;
 
@@ -16,28 +15,19 @@ use ratatui::crossterm::event::KeyModifiers;
 use ratatui::layout::Constraint;
 use ratatui::layout::Direction;
 use ratatui::layout::Layout;
-use ratatui::style::Modifier;
-use ratatui::style::Style;
 use ratatui::text::Line;
 use ratatui::text::Span;
-use ratatui::text::Text;
 use ratatui::widgets::Block;
 use ratatui::widgets::Borders;
 use ratatui::widgets::Paragraph;
-use ratatui::widgets::Scrollbar;
-use ratatui::widgets::ScrollbarOrientation;
-use ratatui::widgets::ScrollbarState;
 use tui_textarea::TextArea;
-use tui_tree_widget::Tree;
 use tui_tree_widget::TreeItem;
 use tui_tree_widget::TreeState;
 
 use crate::library::SnippetLibrary;
 use crate::model::SourceRoot;
 use crate::settings::SettingsCollection;
-use crate::settings::SettingsFile;
 use crate::settings::SettingsLineMap;
-use crate::settings::format_settings_with_map;
 use crate::tui::theme::Theme;
 
 pub type TreeId = String;
@@ -77,7 +67,7 @@ pub struct ContentState {
 }
 
 impl ContentState {
-    fn new() -> Self {
+    pub(crate) fn new() -> Self {
         Self {
             text: None,
             scroll: 0,
@@ -213,7 +203,7 @@ impl SettingsState {
         self.ensure_cursor_visible();
     }
 
-    fn ensure_cursor_visible(&mut self) {
+    pub(crate) fn ensure_cursor_visible(&mut self) {
         // Count visible lines before the cursor to determine effective scroll position.
         let visible_before: usize = (0..=self.cursor)
             .filter(|&i| self.is_line_visible(i))
@@ -336,7 +326,7 @@ impl SettingsState {
 
     /// Returns the visible line count (excluding hidden lines in collapsed
     /// sections).
-    fn visible_line_count(&self) -> usize {
+    pub(crate) fn visible_line_count(&self) -> usize {
         (0..self.lines.len())
             .filter(|&i| self.is_line_visible(i))
             .count()
@@ -352,7 +342,7 @@ pub struct EditState {
     pub discard_confirmed: bool,
     /// Cached dirty flag. `None` means the cache is stale and must be recomputed.
     /// Using `Cell` allows `is_dirty()` to keep `&self` (needed for Debug and draw).
-    dirty_cache: Cell<Option<bool>>,
+    pub(crate) dirty_cache: Cell<Option<bool>>,
 }
 
 impl std::fmt::Debug for EditState {
@@ -388,9 +378,9 @@ pub struct App {
     pub exit: bool,
     pub screen: Screen,
     pub mode: Mode,
-    tree_state: TreeState<TreeId>,
-    tree_items: Vec<TreeItem<'static, TreeId>>,
-    active_pane: Pane,
+    pub(crate) tree_state: TreeState<TreeId>,
+    pub(crate) tree_items: Vec<TreeItem<'static, TreeId>>,
+    pub(crate) active_pane: Pane,
     pub content: ContentState,
     pub title_input: String,
     pub title_cursor: usize,
@@ -462,7 +452,7 @@ impl App {
         Ok(())
     }
 
-    fn help_line(&self) -> Line<'static> {
+    pub(crate) fn help_line(&self) -> Line<'static> {
         let key_style = self.theme.help_key;
         let desc_style = self.theme.help_desc;
         let sep = Span::styled("  ", desc_style);
@@ -552,7 +542,7 @@ impl App {
         Line::from(spans)
     }
 
-    fn draw(&mut self, frame: &mut Frame) {
+    pub(crate) fn draw(&mut self, frame: &mut Frame) {
         // Vertical layout: tab_bar + main area + optional input/status bar + help bar
         let has_input_or_status = self.mode == Mode::TitleInput
             || self.mode == Mode::RenameInput
@@ -635,274 +625,7 @@ impl App {
         frame.render_widget(Paragraph::new(tab_line), area);
     }
 
-    fn draw_files_screen(&mut self, frame: &mut Frame, area: ratatui::layout::Rect) {
-        // In edit mode, render the full area as an editor
-        if self.mode == Mode::Edit {
-            self.draw_edit_pane(frame, area);
-            return;
-        }
-
-        let chunks = Layout::default()
-            .direction(Direction::Horizontal)
-            .constraints([Constraint::Percentage(30), Constraint::Percentage(70)])
-            .split(area);
-
-        let file_border_style = if self.active_pane == Pane::FileList {
-            self.theme.active_border
-        } else {
-            self.theme.inactive_border
-        };
-
-        let content_border_style = if self.active_pane == Pane::Content {
-            self.theme.active_border
-        } else {
-            self.theme.inactive_border
-        };
-
-        if let Ok(tree) = Tree::new(&self.tree_items) {
-            let tree = tree
-                .block(
-                    Block::default()
-                        .borders(Borders::ALL)
-                        .border_style(file_border_style)
-                        .title("CLAUDE.md files"),
-                )
-                .highlight_style(self.theme.highlight);
-            frame.render_stateful_widget(tree, chunks[0], &mut self.tree_state);
-        }
-
-        if self.mode == Mode::LibraryBrowse || self.mode == Mode::RenameInput {
-            self.draw_library_pane(frame, chunks[1], content_border_style);
-        } else {
-            self.draw_content_pane(frame, chunks[1], content_border_style);
-        }
-    }
-
-    fn draw_settings_screen(&mut self, frame: &mut Frame, area: ratatui::layout::Rect) {
-        if self.mode == Mode::Edit {
-            self.draw_edit_pane(frame, area);
-            return;
-        }
-
-        self.settings_state.viewport_height = area.height.saturating_sub(2);
-
-        let cursor_line = self.settings_state.cursor;
-        let cursor_style = self.theme.highlight;
-
-        // Only render visible lines (respecting collapsed sections).
-        let lines: Vec<Line> = self
-            .settings_state
-            .lines
-            .iter()
-            .enumerate()
-            .filter(|&(i, _)| self.settings_state.is_line_visible(i))
-            .map(|(i, line_text)| {
-                let style = if i == cursor_line {
-                    cursor_style
-                } else {
-                    Style::default()
-                };
-                Line::from(line_text.as_str().to_string()).style(style)
-            })
-            .collect();
-
-        let title = if self.settings_state.merged_view {
-            "Settings — Effective"
-        } else {
-            "Settings"
-        };
-        let settings_widget = Paragraph::new(Text::from(lines))
-            .block(Block::default().borders(Borders::ALL).title(title))
-            .scroll((self.settings_state.scroll, 0));
-        frame.render_widget(settings_widget, area);
-
-        let visible_count = self.settings_state.visible_line_count();
-        let mut scrollbar_state =
-            ScrollbarState::new(visible_count).position(self.settings_state.scroll as usize);
-        let scrollbar = Scrollbar::new(ScrollbarOrientation::VerticalRight);
-        frame.render_stateful_widget(scrollbar, area, &mut scrollbar_state);
-    }
-
-    fn draw_content_pane(
-        &mut self,
-        frame: &mut Frame,
-        area: ratatui::layout::Rect,
-        border_style: Style,
-    ) {
-        let content_title = match self.mode {
-            Mode::VisualSelect | Mode::TitleInput => {
-                if let Some((start, end)) = self.content.selection_range() {
-                    format!("Content [VISUAL: lines {}-{}]", start + 1, end + 1)
-                } else {
-                    "Content [VISUAL]".to_string()
-                }
-            }
-            _ => "Content".to_string(),
-        };
-
-        // Capture viewport height (content area minus 2 for borders)
-        self.content.viewport_height = area.height.saturating_sub(2);
-
-        let display_text = self
-            .content
-            .text
-            .as_deref()
-            .unwrap_or("Select a file to view its content.");
-
-        let selection = self.content.selection_range();
-        let cursor_line = self.content.cursor;
-        let show_cursor = self.active_pane == Pane::Content;
-        let cursor_style = self.theme.highlight;
-        let highlight_style = self.theme.visual_selection;
-
-        let lines: Vec<Line> = display_text
-            .lines()
-            .enumerate()
-            .map(|(i, line_text)| {
-                let mut style = Style::default();
-                if let Some((start, end)) = selection
-                    && i >= start
-                    && i <= end
-                {
-                    style = highlight_style;
-                }
-                if show_cursor && i == cursor_line {
-                    style = style.add_modifier(Modifier::REVERSED);
-                    if selection.is_none() {
-                        style = cursor_style;
-                    }
-                }
-                // Ensure the cursor line has at least a space so the
-                // REVERSED style is visible even on empty lines.
-                let text = if show_cursor && i == cursor_line && line_text.is_empty() {
-                    " ".to_string()
-                } else {
-                    line_text.to_string()
-                };
-                Line::from(text).style(style)
-            })
-            .collect();
-
-        let content_widget = Paragraph::new(Text::from(lines))
-            .block(
-                Block::default()
-                    .borders(Borders::ALL)
-                    .border_style(border_style)
-                    .title(content_title),
-            )
-            .scroll((self.content.scroll, 0));
-        frame.render_widget(content_widget, area);
-
-        let mut scrollbar_state =
-            ScrollbarState::new(self.content.line_count()).position(self.content.scroll as usize);
-        let scrollbar = Scrollbar::new(ScrollbarOrientation::VerticalRight);
-        frame.render_stateful_widget(scrollbar, area, &mut scrollbar_state);
-    }
-
-    fn draw_library_pane(
-        &self,
-        frame: &mut Frame,
-        area: ratatui::layout::Rect,
-        border_style: Style,
-    ) {
-        let lib = match &self.library {
-            Some(lib) => lib,
-            None => return,
-        };
-
-        if lib.snippets.is_empty() {
-            let empty_msg = Paragraph::new("No snippets saved. Use v to select, s to save.").block(
-                Block::default()
-                    .borders(Borders::ALL)
-                    .border_style(border_style)
-                    .title("Library (empty)"),
-            );
-            frame.render_widget(empty_msg, area);
-            return;
-        }
-
-        let lib_split = Layout::default()
-            .direction(Direction::Vertical)
-            .constraints([Constraint::Percentage(40), Constraint::Percentage(60)])
-            .split(area);
-
-        // Snippet list (top)
-        let list_title = format!("Library ({} snippets)", lib.snippets.len());
-        let list_lines: Vec<Line> = lib
-            .snippets
-            .iter()
-            .enumerate()
-            .map(|(i, snippet)| {
-                let style = if i == self.library_selected {
-                    self.theme.highlight
-                } else {
-                    Style::default()
-                };
-                Line::from(format!("  {}", snippet.title)).style(style)
-            })
-            .collect();
-        let list_widget = Paragraph::new(Text::from(list_lines)).block(
-            Block::default()
-                .borders(Borders::ALL)
-                .border_style(border_style)
-                .title(list_title),
-        );
-        frame.render_widget(list_widget, lib_split[0]);
-
-        // Preview (bottom)
-        let preview_content = lib
-            .snippets
-            .get(self.library_selected)
-            .map(|s| s.content.as_str())
-            .unwrap_or("");
-        let preview_title = lib
-            .snippets
-            .get(self.library_selected)
-            .map(|s| format!("Preview: {}", s.title))
-            .unwrap_or_else(|| "Preview".to_string());
-        let preview_widget = Paragraph::new(preview_content).block(
-            Block::default()
-                .borders(Borders::ALL)
-                .border_style(border_style)
-                .title(preview_title),
-        );
-        frame.render_widget(preview_widget, lib_split[1]);
-    }
-
-    fn draw_edit_pane(&mut self, frame: &mut Frame, area: ratatui::layout::Rect) {
-        // Use take()/put-back pattern for the mutable borrow
-        let mut edit = match self.edit_state.take() {
-            Some(e) => e,
-            None => return,
-        };
-
-        // Show full path for settings files, just filename for CLAUDE.md
-        let display_name = if self.screen == Screen::Settings {
-            edit.file_path.display().to_string()
-        } else {
-            edit.file_path
-                .file_name()
-                .map(|n| n.to_string_lossy().to_string())
-                .unwrap_or_else(|| edit.file_path.display().to_string())
-        };
-
-        let dirty_marker = if edit.is_dirty() { " [*]" } else { "" };
-        let title = format!("Edit: {display_name}{dirty_marker}");
-
-        edit.textarea.set_block(
-            Block::default()
-                .borders(Borders::ALL)
-                .border_style(self.theme.input_border)
-                .title(title),
-        );
-
-        frame.render_widget(&edit.textarea, area);
-
-        // Put back
-        self.edit_state = Some(edit);
-    }
-
-    fn load_selected_content(&mut self) {
+    pub(crate) fn load_selected_content(&mut self) {
         let selected = self.tree_state.selected();
         if selected.len() < 2 {
             self.content.text = None;
@@ -918,7 +641,7 @@ impl App {
         }
     }
 
-    fn load_file_content(&mut self, path: &Path) {
+    pub(crate) fn load_file_content(&mut self, path: &Path) {
         let text = match fs::read_to_string(path) {
             Ok(text) => text,
             Err(err) => format!("Error reading {}: {err}", path.display()),
@@ -926,315 +649,14 @@ impl App {
         self.content.load_text(text);
     }
 
-    fn reset_to_normal(&mut self) {
+    pub(crate) fn reset_to_normal(&mut self) {
         self.mode = Mode::Normal;
         self.content.visual_anchor = None;
         self.title_input.clear();
         self.title_cursor = 0;
     }
 
-    fn switch_to_settings(&mut self) {
-        let project = std::env::current_dir().unwrap_or_default();
-        self.switch_to_settings_from(&project);
-    }
-
-    /// Switch to settings screen using an explicit project path (for testability).
-    pub fn switch_to_settings_from(&mut self, project: &Path) {
-        let collection = crate::settings::discover_settings_files(project);
-        self.apply_settings_collection(collection);
-        self.screen = Screen::Settings;
-    }
-
-    /// Switch to settings screen with a pre-built collection (for testability).
-    #[cfg(test)]
-    pub fn switch_to_settings_with(&mut self, collection: &SettingsCollection) {
-        self.apply_settings_collection(collection.clone());
-        self.screen = Screen::Settings;
-    }
-
-    fn apply_settings_collection(&mut self, collection: SettingsCollection) {
-        self.settings_collection = Some(collection);
-        self.settings_state.merged_view = false;
-        self.rebuild_settings_display();
-    }
-
-    /// Rebuilds the settings display lines from the cached collection.
-    ///
-    /// Uses per-file formatting or merged formatting depending on
-    /// `settings_state.merged_view`.
-    fn rebuild_settings_display(&mut self) {
-        let Some(collection) = &self.settings_collection else {
-            return;
-        };
-        let (lines, line_map) = if self.settings_state.merged_view {
-            let merged = crate::settings::merge_settings(collection);
-            let synthetic = SettingsCollection {
-                files: vec![SettingsFile {
-                    label: "Effective".to_string(),
-                    path: PathBuf::new(),
-                    value: merged,
-                }],
-            };
-            format_settings_with_map(&synthetic)
-        } else {
-            format_settings_with_map(collection)
-        };
-        self.settings_state.lines = lines;
-        self.settings_state.line_map = line_map;
-        self.settings_state.scroll = 0;
-        self.settings_state.cursor = 0;
-        self.settings_state.collapsed.clear();
-    }
-
-    /// Returns the file path of the settings file at the current cursor position.
-    pub fn settings_file_at_cursor(&self) -> Option<&Path> {
-        let file_idx = self
-            .settings_state
-            .line_map
-            .get(self.settings_state.cursor)
-            .copied()
-            .flatten()?;
-        let collection = self.settings_collection.as_ref()?;
-        let file = collection.files.get(file_idx)?;
-        Some(&file.path)
-    }
-
-    fn enter_edit_mode(&mut self) {
-        let selected = self.tree_state.selected();
-        if selected.len() < 2 {
-            return;
-        }
-        let path_str = match selected.last() {
-            Some(s) => s.clone(),
-            None => return,
-        };
-        let path = PathBuf::from(&path_str);
-        self.enter_edit_mode_for(&path);
-    }
-
-    /// Maximum file size (in bytes) allowed for editing. Files larger than this
-    /// are rejected to prevent excessive memory usage.
-    const MAX_EDIT_FILE_SIZE: u64 = 10 * 1024 * 1024; // 10 MB
-
-    /// Enter edit mode for a specific file path. Extracted for testability.
-    pub fn enter_edit_mode_for(&mut self, path: &Path) {
-        // Guard against very large files to prevent OOM
-        match fs::metadata(path) {
-            Ok(meta) if meta.len() > Self::MAX_EDIT_FILE_SIZE => {
-                let size_mb = meta.len() as f64 / (1024.0 * 1024.0);
-                self.status_message = Some(format!(
-                    "File too large to edit ({size_mb:.1} MB, max 10 MB)"
-                ));
-                return;
-            }
-            Err(err) => {
-                self.status_message = Some(format!("Cannot open for editing: {err}"));
-                return;
-            }
-            _ => {}
-        }
-
-        let raw = match fs::read_to_string(path) {
-            Ok(text) => text,
-            Err(err) => {
-                self.status_message = Some(format!("Cannot open for editing: {err}"));
-                return;
-            }
-        };
-
-        let had_trailing_newline = raw.ends_with('\n');
-        let text = if had_trailing_newline {
-            raw.strip_suffix('\n').unwrap_or(&raw).to_string()
-        } else {
-            raw
-        };
-
-        let lines: Vec<String> = text.lines().map(String::from).collect();
-        let lines = if lines.is_empty() {
-            vec![String::new()]
-        } else {
-            lines
-        };
-
-        let mut textarea = TextArea::new(lines);
-        textarea.set_tab_length(4);
-        textarea.set_cursor_line_style(self.theme.edit_cursor_line);
-
-        self.edit_state = Some(EditState {
-            textarea,
-            file_path: path.to_path_buf(),
-            original_text: text,
-            had_trailing_newline,
-            discard_confirmed: false,
-            dirty_cache: Cell::new(Some(false)),
-        });
-        self.mode = Mode::Edit;
-    }
-
-    fn handle_edit_key(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Char('s') if key_event.modifiers.contains(KeyModifiers::CONTROL) => {
-                self.save_edit();
-            }
-            KeyCode::Esc => {
-                self.exit_edit_mode();
-            }
-            _ => {
-                // Forward all other keys to the textarea
-                if let Some(edit) = &mut self.edit_state {
-                    edit.textarea.input(key_event);
-                    edit.invalidate_dirty_cache();
-                    // Any non-Esc key resets the discard confirmation
-                    edit.discard_confirmed = false;
-                }
-            }
-        }
-    }
-
-    fn save_edit(&mut self) {
-        let Some(edit) = &self.edit_state else {
-            return;
-        };
-        let path = edit.file_path.clone();
-        self.save_edit_to(&path);
-    }
-
-    /// Save the current edit to a specific path. Extracted for testability.
-    pub fn save_edit_to(&mut self, path: &Path) {
-        let Some(edit) = &mut self.edit_state else {
-            return;
-        };
-
-        let joined = edit.textarea.lines().join("\n");
-        let write_content = if edit.had_trailing_newline {
-            format!("{joined}\n")
-        } else {
-            joined.clone()
-        };
-
-        // Atomic write: write to temp file in the same directory, then rename.
-        let parent = path.parent().unwrap_or(Path::new("."));
-        let result = tempfile::NamedTempFile::new_in(parent).and_then(|mut tmp| {
-            tmp.write_all(write_content.as_bytes())?;
-            tmp.flush()?;
-            tmp.persist(path).map_err(|e| e.error)?;
-            Ok(())
-        });
-
-        match result {
-            Ok(()) => {
-                // Update original_text so the dirty flag clears
-                edit.original_text = joined;
-                edit.dirty_cache.set(Some(false));
-                self.status_message = Some("Saved.".to_string());
-            }
-            Err(err) => {
-                self.status_message = Some(format!("Save failed: {err}"));
-            }
-        }
-    }
-
-    fn exit_edit_mode(&mut self) {
-        let Some(edit) = &mut self.edit_state else {
-            self.finalize_exit_edit();
-            return;
-        };
-
-        if edit.is_dirty() && !edit.discard_confirmed {
-            edit.discard_confirmed = true;
-            self.status_message =
-                Some("You have unsaved changes. Press Esc again to discard.".to_string());
-            return;
-        }
-
-        self.finalize_exit_edit();
-    }
-
-    fn finalize_exit_edit(&mut self) {
-        // Reload content into the read-only viewer if on Files screen
-        if self.screen == Screen::Files
-            && let Some(edit) = &self.edit_state
-        {
-            self.load_file_content(&edit.file_path.clone());
-        }
-
-        // If on Settings screen, refresh the formatted view
-        if self.screen == Screen::Settings {
-            self.refresh_settings();
-        }
-
-        self.edit_state = None;
-        self.mode = Mode::Normal;
-    }
-
-    fn refresh_settings(&mut self) {
-        let project = std::env::current_dir().unwrap_or_default();
-        let collection = crate::settings::discover_settings_files(&project);
-        self.apply_settings_collection(collection);
-    }
-
-    fn enter_settings_edit_mode(&mut self) {
-        let path = match self.settings_file_at_cursor() {
-            Some(p) => p.to_path_buf(),
-            None => {
-                self.status_message = Some("No settings file at cursor.".to_string());
-                return;
-            }
-        };
-        self.enter_edit_mode_for(&path);
-    }
-
-    fn handle_settings_key(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Char('e') if !self.settings_state.merged_view => {
-                self.enter_settings_edit_mode();
-            }
-            KeyCode::Char('e') => {
-                self.status_message =
-                    Some("Edit not available in merged view — press m to switch.".to_string());
-            }
-            KeyCode::Char('m') => {
-                self.settings_state.merged_view = !self.settings_state.merged_view;
-                self.rebuild_settings_display();
-            }
-            KeyCode::Char('q') => self.exit = true,
-            KeyCode::Down | KeyCode::Char('j') => {
-                self.settings_state.cursor_down();
-            }
-            KeyCode::Up | KeyCode::Char('k') => {
-                self.settings_state.cursor_up();
-            }
-            KeyCode::PageDown => {
-                self.settings_state.cursor_page_down();
-            }
-            KeyCode::PageUp => {
-                self.settings_state.cursor_page_up();
-            }
-            KeyCode::Left | KeyCode::Char('h') => {
-                let cursor = self.settings_state.cursor;
-                if self.settings_state.is_foldable(cursor)
-                    && !self.settings_state.collapsed.contains(&cursor)
-                {
-                    // On a foldable line: collapse it
-                    self.settings_state.toggle_fold(cursor);
-                } else if let Some(parent) = self.settings_state.parent_for(cursor) {
-                    // On a child line: jump to parent
-                    self.settings_state.cursor = parent;
-                    self.settings_state.ensure_cursor_visible();
-                }
-            }
-            KeyCode::Right | KeyCode::Char('l') => {
-                let cursor = self.settings_state.cursor;
-                if self.settings_state.collapsed.contains(&cursor) {
-                    // On a collapsed line: expand it
-                    self.settings_state.toggle_fold(cursor);
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn current_source_path(&self) -> String {
+    pub(crate) fn current_source_path(&self) -> String {
         self.tree_state
             .selected()
             .last()
@@ -1298,323 +720,6 @@ impl App {
             Screen::Settings => self.handle_settings_key(key_event),
         }
     }
-
-    fn handle_normal_key(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Char('q') => self.exit = true,
-            KeyCode::Tab => {
-                self.active_pane = match self.active_pane {
-                    Pane::FileList => Pane::Content,
-                    Pane::Content => Pane::FileList,
-                };
-            }
-            KeyCode::Down | KeyCode::Char('j') if self.active_pane == Pane::FileList => {
-                self.tree_state.key_down();
-                self.load_selected_content();
-            }
-            KeyCode::Up | KeyCode::Char('k') if self.active_pane == Pane::FileList => {
-                self.tree_state.key_up();
-                self.load_selected_content();
-            }
-            KeyCode::Left | KeyCode::Char('h') if self.active_pane == Pane::FileList => {
-                let before = self.tree_state.selected().to_vec();
-                self.tree_state.key_left();
-                if self.tree_state.selected().is_empty() {
-                    self.tree_state.select(before);
-                }
-                self.load_selected_content();
-            }
-            KeyCode::Right | KeyCode::Char('l') if self.active_pane == Pane::FileList => {
-                self.tree_state.key_right();
-                self.load_selected_content();
-            }
-            KeyCode::Down | KeyCode::Char('j') if self.active_pane == Pane::Content => {
-                self.content.cursor_down();
-            }
-            KeyCode::Up | KeyCode::Char('k') if self.active_pane == Pane::Content => {
-                self.content.cursor_up();
-            }
-            KeyCode::PageDown if self.active_pane == Pane::Content => {
-                self.content.cursor_page_down();
-            }
-            KeyCode::PageUp if self.active_pane == Pane::Content => {
-                self.content.cursor_page_up();
-            }
-            KeyCode::Char('v') if self.active_pane == Pane::Content => {
-                self.content.visual_anchor = Some(self.content.cursor);
-                self.mode = Mode::VisualSelect;
-            }
-            KeyCode::Char('e') if self.active_pane == Pane::Content => {
-                self.enter_edit_mode();
-            }
-            KeyCode::Char('L') if self.active_pane == Pane::Content => {
-                self.enter_library_browse();
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_visual_select_key(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Esc => {
-                self.content.visual_anchor = None;
-                self.mode = Mode::Normal;
-            }
-            KeyCode::Down | KeyCode::Char('j') => {
-                self.content.cursor_down();
-            }
-            KeyCode::Up | KeyCode::Char('k') => {
-                self.content.cursor_up();
-            }
-            KeyCode::Char('s') => {
-                self.title_input.clear();
-                self.title_cursor = 0;
-                self.mode = Mode::TitleInput;
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_title_input_key(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Esc => {
-                self.title_input.clear();
-                self.title_cursor = 0;
-                self.mode = Mode::VisualSelect;
-            }
-            KeyCode::Enter => {
-                self.save_current_snippet();
-            }
-            KeyCode::Backspace => {
-                if self.title_cursor > 0 {
-                    self.title_cursor -= 1;
-                    self.title_input.remove(self.title_cursor);
-                }
-            }
-            KeyCode::Left => {
-                self.title_cursor = self.title_cursor.saturating_sub(1);
-            }
-            KeyCode::Right => {
-                if self.title_cursor < self.title_input.len() {
-                    self.title_cursor += 1;
-                }
-            }
-            KeyCode::Char(c) => {
-                self.title_input.insert(self.title_cursor, c);
-                self.title_cursor += 1;
-            }
-            _ => {}
-        }
-    }
-
-    fn save_current_snippet(&mut self) {
-        match crate::library::library_path() {
-            Some(path) => self.save_current_snippet_to(&path),
-            None => {
-                self.status_message = Some("Cannot determine library path.".to_string());
-                self.reset_to_normal();
-            }
-        }
-    }
-
-    /// Save snippet to a specific path. Extracted for testability.
-    pub fn save_current_snippet_to(&mut self, path: &Path) {
-        let title = self.title_input.trim().to_string();
-        if title.is_empty() {
-            self.status_message = Some("Title cannot be empty.".to_string());
-            return;
-        }
-
-        let selected_text = match self.content.selected_text() {
-            Some(text) => text,
-            None => {
-                self.status_message = Some("No text selected.".to_string());
-                self.reset_to_normal();
-                return;
-            }
-        };
-
-        let source = self.current_source_path();
-
-        let snippet = crate::library::Snippet {
-            title,
-            content: selected_text,
-            source,
-        };
-
-        match crate::library::append_snippet(snippet, path) {
-            Ok(()) => {
-                self.status_message = Some("Snippet saved!".to_string());
-            }
-            Err(err) => {
-                self.status_message = Some(format!("Save failed: {err}"));
-            }
-        }
-
-        self.reset_to_normal();
-    }
-
-    fn enter_library_browse(&mut self) {
-        match crate::library::library_path() {
-            Some(path) => self.enter_library_browse_from(&path),
-            None => {
-                self.status_message = Some("Cannot determine library path.".to_string());
-            }
-        }
-    }
-
-    /// Enter library browse with a specific path. Extracted for testability.
-    pub fn enter_library_browse_from(&mut self, path: &Path) {
-        match crate::library::load_library(path) {
-            Ok(lib) => {
-                self.library = Some(lib);
-                self.library_selected = 0;
-                self.mode = Mode::LibraryBrowse;
-            }
-            Err(err) => {
-                self.status_message = Some(format!("Failed to load library: {err}"));
-            }
-        }
-    }
-
-    fn handle_library_browse_key(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Esc | KeyCode::Char('q') => {
-                self.library = None;
-                self.mode = Mode::Normal;
-            }
-            KeyCode::Down | KeyCode::Char('j') => {
-                let max = self
-                    .library
-                    .as_ref()
-                    .map_or(0, |lib| lib.snippets.len().saturating_sub(1));
-                if self.library_selected < max {
-                    self.library_selected += 1;
-                }
-            }
-            KeyCode::Up | KeyCode::Char('k') => {
-                self.library_selected = self.library_selected.saturating_sub(1);
-            }
-            KeyCode::Char('d') => {
-                self.delete_library_snippet();
-            }
-            KeyCode::Char('r') => {
-                if let Some(lib) = &self.library
-                    && let Some(snippet) = lib.snippets.get(self.library_selected)
-                {
-                    self.title_input = snippet.title.clone();
-                    self.title_cursor = self.title_input.len();
-                    self.mode = Mode::RenameInput;
-                }
-            }
-            _ => {}
-        }
-    }
-
-    fn handle_rename_input_key(&mut self, key_event: KeyEvent) {
-        match key_event.code {
-            KeyCode::Esc => {
-                self.title_input.clear();
-                self.title_cursor = 0;
-                self.mode = Mode::LibraryBrowse;
-            }
-            KeyCode::Enter => {
-                self.rename_library_snippet();
-            }
-            KeyCode::Backspace => {
-                if self.title_cursor > 0 {
-                    self.title_cursor -= 1;
-                    self.title_input.remove(self.title_cursor);
-                }
-            }
-            KeyCode::Left => {
-                self.title_cursor = self.title_cursor.saturating_sub(1);
-            }
-            KeyCode::Right => {
-                if self.title_cursor < self.title_input.len() {
-                    self.title_cursor += 1;
-                }
-            }
-            KeyCode::Char(c) => {
-                self.title_input.insert(self.title_cursor, c);
-                self.title_cursor += 1;
-            }
-            _ => {}
-        }
-    }
-
-    fn rename_library_snippet(&mut self) {
-        match crate::library::library_path() {
-            Some(path) => self.rename_library_snippet_from(&path),
-            None => {
-                self.status_message = Some("Cannot determine library path.".to_string());
-                self.title_input.clear();
-                self.title_cursor = 0;
-                self.mode = Mode::LibraryBrowse;
-            }
-        }
-    }
-
-    /// Rename a library snippet. Extracted for testability.
-    pub fn rename_library_snippet_from(&mut self, path: &Path) {
-        let new_title = self.title_input.trim().to_string();
-        if new_title.is_empty() {
-            self.status_message = Some("Title cannot be empty.".to_string());
-            return;
-        }
-
-        match crate::library::rename_snippet(self.library_selected, &new_title, path) {
-            Ok(()) => {
-                if let Ok(lib) = crate::library::load_library(path) {
-                    self.library = Some(lib);
-                }
-                self.status_message = Some("Snippet renamed.".to_string());
-            }
-            Err(err) => {
-                self.status_message = Some(format!("Rename failed: {err}"));
-            }
-        }
-
-        self.title_input.clear();
-        self.title_cursor = 0;
-        self.mode = Mode::LibraryBrowse;
-    }
-
-    fn delete_library_snippet(&mut self) {
-        match crate::library::library_path() {
-            Some(path) => self.delete_library_snippet_from(&path),
-            None => {
-                self.status_message = Some("Cannot determine library path.".to_string());
-            }
-        }
-    }
-
-    /// Delete a library snippet at a specific path. Extracted for testability.
-    pub fn delete_library_snippet_from(&mut self, path: &Path) {
-        let snippet_count = self.library.as_ref().map_or(0, |lib| lib.snippets.len());
-        if snippet_count == 0 {
-            return;
-        }
-
-        match crate::library::delete_snippet(self.library_selected, path) {
-            Ok(()) => {
-                // Reload library from disk
-                if let Ok(lib) = crate::library::load_library(path) {
-                    let new_len = lib.snippets.len();
-                    self.library = Some(lib);
-                    if self.library_selected >= new_len && new_len > 0 {
-                        self.library_selected = new_len - 1;
-                    } else if new_len == 0 {
-                        self.library_selected = 0;
-                    }
-                }
-                self.status_message = Some("Snippet deleted.".to_string());
-            }
-            Err(err) => {
-                self.status_message = Some(format!("Delete failed: {err}"));
-            }
-        }
-    }
 }
 
 pub fn build_tree_items(roots: &[SourceRoot]) -> Vec<TreeItem<'static, TreeId>> {
@@ -1641,16 +746,22 @@ pub fn build_tree_items(roots: &[SourceRoot]) -> Vec<TreeItem<'static, TreeId>> 
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::Config;
+pub(crate) mod test_helpers {
+    use std::path::PathBuf;
+
     use ratatui::Terminal;
     use ratatui::backend::TestBackend;
+    use ratatui::crossterm::event::KeyCode;
+    use ratatui::crossterm::event::KeyEvent;
     use ratatui::crossterm::event::KeyEventKind;
     use ratatui::crossterm::event::KeyEventState;
-    use tempfile::TempDir;
+    use ratatui::crossterm::event::KeyModifiers;
 
-    fn key_event(code: KeyCode) -> KeyEvent {
+    use crate::model::SourceRoot;
+
+    use super::App;
+
+    pub fn key_event(code: KeyCode) -> KeyEvent {
         KeyEvent {
             code,
             modifiers: KeyModifiers::empty(),
@@ -1659,7 +770,7 @@ mod tests {
         }
     }
 
-    fn sample_roots() -> Vec<SourceRoot> {
+    pub fn sample_roots() -> Vec<SourceRoot> {
         vec![
             SourceRoot {
                 path: PathBuf::from("/a"),
@@ -1674,6 +785,24 @@ mod tests {
             },
         ]
     }
+
+    /// Renders the app once so `TreeState` populates `last_identifiers`,
+    /// enabling `key_down()`/`key_up()` navigation in tests.
+    pub fn render_once(app: &mut App) {
+        let backend = TestBackend::new(80, 24);
+        let mut terminal = Terminal::new(backend).unwrap();
+        terminal.draw(|frame| app.draw(frame)).unwrap();
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::test_helpers::*;
+    use super::*;
+    use crate::config::Config;
+    use ratatui::crossterm::event::KeyEventKind;
+    use ratatui::crossterm::event::KeyEventState;
+    use tempfile::TempDir;
 
     #[test]
     fn q_key_sets_exit() {
@@ -1700,252 +829,6 @@ mod tests {
     }
 
     #[test]
-    fn tab_toggles_pane() {
-        let mut app = App::new(sample_roots(), &Config::default());
-        assert_eq!(app.active_pane, Pane::FileList);
-
-        app.handle_key_event(key_event(KeyCode::Tab));
-        assert_eq!(app.active_pane, Pane::Content);
-
-        app.handle_key_event(key_event(KeyCode::Tab));
-        assert_eq!(app.active_pane, Pane::FileList);
-    }
-
-    #[test]
-    fn arrow_keys_ignored_when_content_pane_active() {
-        let mut app = App::new(sample_roots(), &Config::default());
-        let initial_selected = app.tree_state.selected().to_vec();
-
-        app.handle_key_event(key_event(KeyCode::Tab));
-        assert_eq!(app.active_pane, Pane::Content);
-
-        app.handle_key_event(key_event(KeyCode::Down));
-        assert_eq!(app.tree_state.selected(), initial_selected);
-    }
-
-    /// Render the app once so TreeState populates `last_identifiers`,
-    /// enabling `key_down()`/`key_up()` navigation in tests.
-    fn render_once(app: &mut App) {
-        let backend = TestBackend::new(80, 24);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal.draw(|frame| app.draw(frame)).unwrap();
-    }
-
-    #[test]
-    fn jk_can_land_on_folder_node() {
-        let mut app = App::new(sample_roots(), &Config::default());
-        render_once(&mut app);
-
-        // App starts on first file /a/CLAUDE.md — selected len is 2
-        assert_eq!(app.tree_state.selected().len(), 2);
-
-        // Press k (up) — should land on the /a folder node (len 1)
-        app.handle_key_event(key_event(KeyCode::Char('k')));
-        assert_eq!(
-            app.tree_state.selected().len(),
-            1,
-            "k should be able to land on a folder node"
-        );
-    }
-
-    #[test]
-    fn folder_selection_clears_content_pane() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Some content").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file],
-        }];
-        let mut app = App::new(roots, &Config::default());
-
-        // Content is loaded on startup
-        assert!(app.content.text.is_some());
-
-        // Select the root/folder node
-        app.tree_state
-            .select(vec![tmp.path().display().to_string()]);
-        app.load_selected_content();
-
-        assert!(
-            app.content.text.is_none(),
-            "Content pane should be cleared when a folder is selected"
-        );
-    }
-
-    #[test]
-    fn left_arrow_to_parent_clears_content() {
-        let mut app = App::new(sample_roots(), &Config::default());
-        render_once(&mut app);
-
-        // Start on first file — content is loaded
-        assert_eq!(app.tree_state.selected().len(), 2);
-        assert!(app.content.text.is_some());
-
-        // Press Left — should navigate to parent folder
-        app.handle_key_event(key_event(KeyCode::Left));
-
-        assert_eq!(
-            app.tree_state.selected().len(),
-            1,
-            "Left should navigate to parent folder"
-        );
-        assert!(
-            app.content.text.is_none(),
-            "Content should be cleared when folder is selected via Left"
-        );
-    }
-
-    #[test]
-    fn left_on_folder_node_does_not_lose_selection() {
-        let mut app = App::new(sample_roots(), &Config::default());
-        render_once(&mut app);
-
-        // Navigate to the /a folder node
-        app.tree_state.select(vec!["/a".to_string()]);
-        assert_eq!(app.tree_state.selected().len(), 1);
-
-        // First Left closes the folder (it starts open) — stays on folder
-        app.handle_key_event(key_event(KeyCode::Left));
-        assert_eq!(
-            app.tree_state.selected().len(),
-            1,
-            "First Left should close folder, selection stays"
-        );
-
-        // Second Left on a closed folder — selection must not become empty
-        app.handle_key_event(key_event(KeyCode::Left));
-        assert!(
-            !app.tree_state.selected().is_empty(),
-            "Second Left on a closed folder should not clear the selection"
-        );
-        assert_eq!(
-            app.tree_state.selected().len(),
-            1,
-            "Selection should remain on the folder node"
-        );
-    }
-
-    #[test]
-    fn cursor_on_empty_line_is_visible() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        // File with an empty second line
-        fs::write(&file, "first\n\nthird").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file],
-        }];
-        let mut app = App::new(roots, &Config::default());
-        app.active_pane = Pane::Content;
-
-        // Move cursor to the empty line (line index 1)
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.content.cursor, 1);
-
-        let backend = TestBackend::new(80, 24);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal.draw(|frame| app.draw(frame)).unwrap();
-        let buf = terminal.backend().buffer().clone();
-
-        // The cursor line is row 3 in the buffer (row 0 = tab bar, row 1 = border,
-        // row 2 = first content line, row 3 = empty cursor line).
-        // Check that the empty line has a non-default style (Reversed modifier).
-        let content_x_start = (80u16 * 30 / 100) + 1;
-        let cell = &buf[(content_x_start, 3)];
-        assert!(
-            cell.modifier.contains(Modifier::REVERSED),
-            "Empty cursor line should use REVERSED style for visibility, got: {:?}",
-            cell.modifier
-        );
-    }
-
-    #[test]
-    fn enter_in_file_list_is_noop() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Test content").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file.clone()],
-        }];
-        let mut app = App::new(roots, &Config::default());
-
-        // Snapshot state before pressing Enter
-        let pane_before = app.active_pane;
-        let mode_before = app.mode;
-        let content_before = app.content.text.clone();
-        let selected_before = app.tree_state.selected().to_vec();
-
-        // Press Enter on a file node — should be a no-op
-        app.handle_key_event(key_event(KeyCode::Enter));
-
-        assert_eq!(
-            app.active_pane, pane_before,
-            "Enter should not change active pane"
-        );
-        assert_eq!(app.mode, mode_before, "Enter should not change mode");
-        assert_eq!(
-            app.content.text, content_before,
-            "Enter should not reload content"
-        );
-        assert_eq!(
-            app.tree_state.selected(),
-            selected_before,
-            "Enter should not change selection"
-        );
-    }
-
-    #[test]
-    fn enter_on_root_node_is_noop() {
-        let mut app = App::new(sample_roots(), &Config::default());
-
-        // Select a root node
-        app.tree_state.select(vec!["/a".to_string()]);
-        let opened_before = app.tree_state.opened().clone();
-
-        // Press Enter — should not toggle the folder
-        app.handle_key_event(key_event(KeyCode::Enter));
-
-        assert_eq!(
-            app.tree_state.opened().clone(),
-            opened_before,
-            "Enter should not toggle folder open/closed"
-        );
-    }
-
-    #[test]
-    fn toggle_selected_on_root_toggles() {
-        let mut app = App::new(sample_roots(), &Config::default());
-
-        // Directly select a root node (single-segment identifier)
-        app.tree_state.select(vec!["/a".to_string()]);
-
-        let initially_opened = app.tree_state.opened().clone();
-        assert!(
-            initially_opened.contains(&vec!["/a".to_string()]),
-            "Root /a should be open initially"
-        );
-
-        // Toggle via tree_state directly — should close
-        app.tree_state.toggle_selected();
-        assert!(
-            !app.tree_state.opened().contains(&vec!["/a".to_string()]),
-            "Root /a should be closed after toggle"
-        );
-
-        // Toggle again — should open
-        app.tree_state.toggle_selected();
-        assert!(
-            app.tree_state.opened().contains(&vec!["/a".to_string()]),
-            "Root /a should be open after second toggle"
-        );
-    }
-
-    #[test]
     fn first_file_is_selected_and_loaded_on_startup() {
         let tmp = TempDir::new().unwrap();
         let file = tmp.path().join("CLAUDE.md");
@@ -1963,287 +846,6 @@ mod tests {
             app.tree_state.selected(),
             vec![tmp.path().display().to_string(), file.display().to_string()]
         );
-    }
-
-    #[test]
-    fn load_selected_content_loads_file() {
-        let tmp = TempDir::new().unwrap();
-
-        let file_a = tmp.path().join("CLAUDE.md");
-        fs::write(&file_a, "First content").unwrap();
-
-        let sub = tmp.path().join("sub");
-        fs::create_dir_all(&sub).unwrap();
-        let file_b = sub.join("CLAUDE.md");
-        fs::write(&file_b, "Second content").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file_a, file_b.clone()],
-        }];
-        let mut app = App::new(roots, &Config::default());
-
-        // First file is loaded on startup
-        assert_eq!(app.content.text.as_deref(), Some("First content"));
-
-        // Select a different file and load content directly
-        app.tree_state.select(vec![
-            tmp.path().display().to_string(),
-            file_b.display().to_string(),
-        ]);
-        app.load_selected_content();
-        assert_eq!(app.content.text.as_deref(), Some("Second content"));
-    }
-
-    #[test]
-    fn load_content_handles_missing_file() {
-        let roots = vec![SourceRoot {
-            path: PathBuf::from("/nonexistent"),
-            files: vec![PathBuf::from("/nonexistent/CLAUDE.md")],
-        }];
-        let mut app = App::new(roots, &Config::default());
-
-        // Directly select the file node and load content
-        app.tree_state.select(vec![
-            "/nonexistent".to_string(),
-            "/nonexistent/CLAUDE.md".to_string(),
-        ]);
-        app.load_selected_content();
-        assert!(app.content.text.is_some());
-        assert!(
-            app.content
-                .text
-                .as_deref()
-                .unwrap()
-                .contains("Error reading")
-        );
-    }
-
-    #[test]
-    fn cursor_moves_down_and_scrolls_when_past_viewport() {
-        let mut app = App::new(vec![], &Config::default());
-        app.content.text = Some("Line 0\nLine 1\nLine 2\nLine 3\nLine 4".to_string());
-        app.content.viewport_height = 3; // can see 3 lines
-        app.active_pane = Pane::Content;
-
-        app.handle_key_event(key_event(KeyCode::Down));
-        assert_eq!(app.content.cursor, 1);
-        assert_eq!(app.content.scroll, 0, "Still visible, no scroll");
-
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.content.cursor, 2);
-        assert_eq!(app.content.scroll, 0, "Line 2 is last visible row");
-
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.content.cursor, 3);
-        assert_eq!(app.content.scroll, 1, "Scrolls to keep cursor visible");
-    }
-
-    #[test]
-    fn cursor_does_not_go_below_zero() {
-        let mut app = App::new(vec![], &Config::default());
-        app.content.text = Some("Line 0\nLine 1".to_string());
-        app.active_pane = Pane::Content;
-
-        app.handle_key_event(key_event(KeyCode::Up));
-        assert_eq!(app.content.cursor, 0);
-    }
-
-    #[test]
-    fn cursor_clamps_at_last_line() {
-        let mut app = App::new(vec![], &Config::default());
-        app.content.text = Some("Line 0\nLine 1\nLine 2\nLine 3\nLine 4".to_string());
-        app.content.viewport_height = 3;
-        app.active_pane = Pane::Content;
-
-        app.handle_key_event(key_event(KeyCode::PageDown));
-        assert_eq!(app.content.cursor, 3, "Page down moves by viewport_height");
-
-        app.handle_key_event(key_event(KeyCode::PageDown));
-        assert_eq!(app.content.cursor, 4, "Clamps at last line");
-    }
-
-    #[test]
-    fn loading_new_content_resets_scroll_and_cursor() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Line 0\nLine 1\nLine 2").unwrap();
-
-        let root_id = tmp.path().display().to_string();
-        let file_id = file.display().to_string();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file],
-        }];
-        let mut app = App::new(roots, &Config::default());
-
-        // Manually set scroll and cursor
-        app.content.scroll = 5;
-        app.content.cursor = 5;
-
-        // Load file content directly
-        app.tree_state
-            .select(vec![root_id.clone(), file_id.clone()]);
-        app.load_selected_content();
-        assert_eq!(app.content.scroll, 0, "Loading new content resets scroll");
-        assert_eq!(app.content.cursor, 0, "Loading new content resets cursor");
-    }
-
-    /// Extract the first content row text from the content pane in the rendered buffer.
-    fn extract_content_first_line(buf: &ratatui::buffer::Buffer, width: u16) -> String {
-        // Row 0 = tab bar, row 1 = border top of content pane,
-        // row 2 = first content line inside the border.
-        let content_x_start = (width * 30 / 100) + 1;
-        let content_x_end = width - 1; // exclude right border
-        (content_x_start..content_x_end)
-            .map(|x| buf[(x, 2)].symbol().to_string())
-            .collect::<String>()
-    }
-
-    #[test]
-    fn switching_files_does_not_leave_leftover_characters() {
-        let tmp = TempDir::new().unwrap();
-
-        // First file has a long first line
-        let dir_a = tmp.path().join("a");
-        fs::create_dir_all(&dir_a).unwrap();
-        let file_a = dir_a.join("CLAUDE.md");
-        fs::write(&file_a, "# CLAUDE.md\nSecond line").unwrap();
-
-        // Second file has a shorter first line
-        let dir_b = tmp.path().join("b");
-        fs::create_dir_all(&dir_b).unwrap();
-        let file_b = dir_b.join("CLAUDE.md");
-        fs::write(&file_b, "# Short\nOther").unwrap();
-
-        let roots = vec![
-            SourceRoot {
-                path: dir_a.clone(),
-                files: vec![file_a.clone()],
-            },
-            SourceRoot {
-                path: dir_b.clone(),
-                files: vec![file_b.clone()],
-            },
-        ];
-        let mut app = App::new(roots, &Config::default());
-        let width: u16 = 80;
-        let height: u16 = 10;
-
-        let backend = TestBackend::new(width, height);
-        let mut terminal = Terminal::new(backend).unwrap();
-
-        // Draw 1: placeholder
-        terminal.draw(|frame| app.draw(frame)).unwrap();
-
-        // Load the long file and draw
-        app.tree_state.select(vec![
-            dir_a.display().to_string(),
-            file_a.display().to_string(),
-        ]);
-        app.load_selected_content();
-        terminal.draw(|frame| app.draw(frame)).unwrap();
-
-        let buf = terminal.backend().buffer().clone();
-        let line = extract_content_first_line(&buf, width);
-        assert_eq!(
-            line.trim_end(),
-            "# CLAUDE.md",
-            "Long file should render correctly"
-        );
-
-        // Now switch to the shorter file and draw
-        app.tree_state.select(vec![
-            dir_b.display().to_string(),
-            file_b.display().to_string(),
-        ]);
-        app.load_selected_content();
-        terminal.draw(|frame| app.draw(frame)).unwrap();
-
-        let buf = terminal.backend().buffer().clone();
-        let line = extract_content_first_line(&buf, width);
-        eprintln!("RAW content row after Draw 3 (# Short): '{line}'");
-
-        // Also check the Terminal's internal buffer directly for comparison
-        // The TestBackend buffer should match the screen output
-        eprintln!("TestBackend buf cell symbols at row 2, x=25..40:");
-        for x in 25u16..40 {
-            let sym = buf[(x, 2)].symbol();
-            eprint!("[{x}:{}]", sym.escape_debug());
-        }
-        eprintln!();
-
-        let trimmed = line.trim_end();
-
-        assert_eq!(
-            trimmed, "# Short",
-            "After switching to shorter file, first line must not have leftover chars. Got: '{trimmed}'"
-        );
-    }
-
-    #[test]
-    fn tabs_are_expanded_to_spaces() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "\tindented\n\t\tdouble").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file.clone()],
-        }];
-        let mut app = App::new(roots, &Config::default());
-
-        let root_id = tmp.path().display().to_string();
-        let file_id = file.display().to_string();
-        app.tree_state.select(vec![root_id, file_id]);
-        app.load_selected_content();
-
-        let content = app.content.text.as_deref().unwrap();
-        assert!(
-            !content.contains('\t'),
-            "Tabs should be replaced with spaces, got: {content:?}"
-        );
-        assert!(content.starts_with("    indented"));
-    }
-
-    // --- ContentState unit tests ---
-
-    #[test]
-    fn content_state_selection_range_returns_none_without_anchor() {
-        let state = ContentState::new();
-        assert_eq!(state.selection_range(), None);
-    }
-
-    #[test]
-    fn content_state_selection_range_sorts_anchor_and_cursor() {
-        let mut state = ContentState::new();
-        state.visual_anchor = Some(5);
-        state.cursor = 2;
-        assert_eq!(state.selection_range(), Some((2, 5)));
-
-        state.cursor = 8;
-        assert_eq!(state.selection_range(), Some((5, 8)));
-    }
-
-    #[test]
-    fn content_state_selected_text_extracts_lines() {
-        let mut state = ContentState::new();
-        state.text = Some("line 0\nline 1\nline 2\nline 3\nline 4".to_string());
-        state.visual_anchor = Some(1);
-        state.cursor = 3;
-
-        assert_eq!(
-            state.selected_text(),
-            Some("line 1\nline 2\nline 3".to_string())
-        );
-    }
-
-    #[test]
-    fn content_state_selected_text_returns_none_without_anchor() {
-        let mut state = ContentState::new();
-        state.text = Some("line 0\nline 1".to_string());
-        assert_eq!(state.selected_text(), None);
     }
 
     // --- Mode transition tests ---
@@ -2300,801 +902,6 @@ mod tests {
         assert!(app.status_message.is_none());
     }
 
-    // --- Visual selection integration tests ---
-
-    #[test]
-    fn v_in_content_pane_enters_visual_select() {
-        let mut app = App::new(vec![], &Config::default());
-        app.content.text = Some("line 0\nline 1\nline 2".to_string());
-        app.active_pane = Pane::Content;
-        app.content.cursor = 1;
-
-        app.handle_key_event(key_event(KeyCode::Char('v')));
-
-        assert_eq!(app.mode, Mode::VisualSelect);
-        assert_eq!(app.content.visual_anchor, Some(1));
-    }
-
-    #[test]
-    fn v_in_file_list_does_not_enter_visual_select() {
-        let mut app = App::new(vec![], &Config::default());
-        app.active_pane = Pane::FileList;
-
-        app.handle_key_event(key_event(KeyCode::Char('v')));
-
-        assert_eq!(app.mode, Mode::Normal);
-    }
-
-    #[test]
-    fn esc_in_visual_select_returns_to_normal() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::VisualSelect;
-        app.content.visual_anchor = Some(3);
-
-        app.handle_key_event(key_event(KeyCode::Esc));
-
-        assert_eq!(app.mode, Mode::Normal);
-        assert_eq!(app.content.visual_anchor, None);
-    }
-
-    #[test]
-    fn jk_in_visual_select_moves_cursor() {
-        let mut app = App::new(vec![], &Config::default());
-        app.content.text = Some("line 0\nline 1\nline 2\nline 3\nline 4".to_string());
-        app.content.viewport_height = 10;
-        app.mode = Mode::VisualSelect;
-        app.content.visual_anchor = Some(1);
-        app.content.cursor = 1;
-
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.content.cursor, 2);
-        assert_eq!(app.content.selection_range(), Some((1, 2)));
-
-        app.handle_key_event(key_event(KeyCode::Char('k')));
-        assert_eq!(app.content.cursor, 1);
-        assert_eq!(app.content.selection_range(), Some((1, 1)));
-    }
-
-    #[test]
-    fn s_in_visual_select_enters_title_input() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::VisualSelect;
-        app.content.visual_anchor = Some(0);
-
-        app.handle_key_event(key_event(KeyCode::Char('s')));
-
-        assert_eq!(app.mode, Mode::TitleInput);
-        assert!(app.title_input.is_empty());
-    }
-
-    #[test]
-    fn loading_new_content_clears_visual_anchor() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "content").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file.clone()],
-        }];
-        let mut app = App::new(roots, &Config::default());
-        app.content.visual_anchor = Some(5);
-
-        // Re-load the same file
-        let root_id = tmp.path().display().to_string();
-        let file_id = file.display().to_string();
-        app.tree_state.select(vec![root_id, file_id]);
-        app.load_selected_content();
-
-        assert_eq!(app.content.visual_anchor, None);
-    }
-
-    // --- Title input integration tests ---
-
-    #[test]
-    fn title_input_chars_accumulate() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::TitleInput;
-
-        app.handle_key_event(key_event(KeyCode::Char('A')));
-        app.handle_key_event(key_event(KeyCode::Char('B')));
-        assert_eq!(app.title_input, "AB");
-    }
-
-    #[test]
-    fn title_input_backspace_deletes_at_cursor() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::TitleInput;
-        app.title_input = "ABC".to_string();
-        app.title_cursor = 3;
-
-        app.handle_key_event(key_event(KeyCode::Backspace));
-        assert_eq!(app.title_input, "AB");
-        assert_eq!(app.title_cursor, 2);
-    }
-
-    #[test]
-    fn title_input_esc_returns_to_visual_select() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::TitleInput;
-        app.content.visual_anchor = Some(2);
-        app.title_input = "partial".to_string();
-
-        app.handle_key_event(key_event(KeyCode::Esc));
-
-        assert_eq!(app.mode, Mode::VisualSelect);
-        assert_eq!(app.content.visual_anchor, Some(2), "Selection preserved");
-        assert!(app.title_input.is_empty(), "Input cleared on Esc");
-    }
-
-    #[test]
-    fn save_with_empty_title_shows_error() {
-        let tmp = TempDir::new().unwrap();
-        let library_path = tmp.path().join("library.toml");
-
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::TitleInput;
-        app.title_input = "  ".to_string();
-
-        app.save_current_snippet_to(&library_path);
-
-        assert_eq!(app.mode, Mode::TitleInput, "Stays in TitleInput on empty");
-        assert!(app.status_message.as_deref().unwrap().contains("empty"),);
-    }
-
-    #[test]
-    fn title_input_enter_saves_snippet_to_disk() {
-        let tmp = TempDir::new().unwrap();
-        let library_path = tmp.path().join("library.toml");
-
-        let mut app = App::new(vec![], &Config::default());
-        app.content.text = Some("line 0\nline 1\nline 2\nline 3".to_string());
-        app.content.visual_anchor = Some(1);
-        app.content.cursor = 2;
-        app.mode = Mode::TitleInput;
-        app.title_input = "My Snippet".to_string();
-
-        // We can't easily override library_path() in tests, so test the
-        // underlying logic via save_current_snippet_to().
-        app.save_current_snippet_to(&library_path);
-
-        assert_eq!(app.mode, Mode::Normal);
-        assert_eq!(app.content.visual_anchor, None);
-        assert!(app.title_input.is_empty());
-        assert!(app.status_message.as_deref().unwrap().contains("saved"),);
-
-        // Verify the file was written
-        let lib = crate::library::load_library(&library_path).unwrap();
-        assert_eq!(lib.snippets.len(), 1);
-        assert_eq!(lib.snippets[0].title, "My Snippet");
-        assert_eq!(lib.snippets[0].content, "line 1\nline 2");
-    }
-
-    #[test]
-    fn full_visual_select_to_save_flow() {
-        let tmp_content = TempDir::new().unwrap();
-        let file = tmp_content.path().join("CLAUDE.md");
-        fs::write(&file, "# Rules\n- Rule A\n- Rule B\n- Rule C").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp_content.path().to_path_buf(),
-            files: vec![file],
-        }];
-        let mut app = App::new(roots, &Config::default());
-
-        // Switch to content pane
-        app.handle_key_event(key_event(KeyCode::Tab));
-        assert_eq!(app.active_pane, Pane::Content);
-        assert_eq!(app.mode, Mode::Normal);
-
-        // Start visual selection at line 0 (scroll = 0)
-        app.handle_key_event(key_event(KeyCode::Char('v')));
-        assert_eq!(app.mode, Mode::VisualSelect);
-        assert_eq!(app.content.visual_anchor, Some(0));
-
-        // Scroll down two lines
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.content.selection_range(), Some((0, 2)));
-
-        // Press s to enter title input
-        app.handle_key_event(key_event(KeyCode::Char('s')));
-        assert_eq!(app.mode, Mode::TitleInput);
-
-        // Type a title
-        for c in "My Rules".chars() {
-            app.handle_key_event(key_event(KeyCode::Char(c)));
-        }
-        assert_eq!(app.title_input, "My Rules");
-
-        // Save to a temp library path
-        let tmp_lib = TempDir::new().unwrap();
-        let library_path = tmp_lib.path().join("library.toml");
-        app.save_current_snippet_to(&library_path);
-
-        assert_eq!(app.mode, Mode::Normal);
-
-        let lib = crate::library::load_library(&library_path).unwrap();
-        assert_eq!(lib.snippets.len(), 1);
-        assert_eq!(lib.snippets[0].title, "My Rules");
-        assert_eq!(lib.snippets[0].content, "# Rules\n- Rule A\n- Rule B");
-    }
-
-    // --- Library browse tests ---
-
-    fn library_with_snippets(path: &std::path::Path, titles: &[&str]) {
-        for title in titles {
-            crate::library::append_snippet(
-                crate::library::Snippet {
-                    title: title.to_string(),
-                    content: format!("Content of {title}"),
-                    source: "/test/CLAUDE.md".to_string(),
-                },
-                path,
-            )
-            .unwrap();
-        }
-    }
-
-    #[test]
-    fn l_in_content_pane_enters_library_browse() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["Snippet A"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.active_pane = Pane::Content;
-        app.enter_library_browse_from(&lib_path);
-
-        assert_eq!(app.mode, Mode::LibraryBrowse);
-        assert_eq!(app.library_selected, 0);
-        assert!(app.library.is_some());
-        assert_eq!(app.library.as_ref().unwrap().snippets.len(), 1);
-    }
-
-    #[test]
-    fn l_in_file_list_does_not_enter_library_browse() {
-        let mut app = App::new(vec![], &Config::default());
-        app.active_pane = Pane::FileList;
-
-        app.handle_key_event(key_event(KeyCode::Char('L')));
-
-        assert_eq!(app.mode, Mode::Normal);
-    }
-
-    #[test]
-    fn esc_in_library_browse_returns_to_normal() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::LibraryBrowse;
-        app.library = Some(crate::library::SnippetLibrary::default());
-
-        app.handle_key_event(key_event(KeyCode::Esc));
-
-        assert_eq!(app.mode, Mode::Normal);
-        assert!(app.library.is_none(), "Library freed on exit");
-    }
-
-    #[test]
-    fn q_in_library_browse_returns_to_normal_not_exit() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::LibraryBrowse;
-        app.library = Some(crate::library::SnippetLibrary::default());
-
-        app.handle_key_event(key_event(KeyCode::Char('q')));
-
-        assert_eq!(app.mode, Mode::Normal);
-        assert!(!app.exit, "q should not exit the app from LibraryBrowse");
-    }
-
-    #[test]
-    fn jk_in_library_browse_navigates() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["A", "B", "C"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-        assert_eq!(app.library_selected, 0);
-
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.library_selected, 1);
-
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.library_selected, 2);
-
-        // Clamp at end
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.library_selected, 2);
-
-        app.handle_key_event(key_event(KeyCode::Char('k')));
-        assert_eq!(app.library_selected, 1);
-
-        // Clamp at start
-        app.handle_key_event(key_event(KeyCode::Char('k')));
-        app.handle_key_event(key_event(KeyCode::Char('k')));
-        assert_eq!(app.library_selected, 0);
-    }
-
-    #[test]
-    fn d_in_library_browse_deletes_snippet() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["A", "B", "C"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-
-        // Select "B" (index 1) and delete it
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.library_selected, 1);
-
-        app.delete_library_snippet_from(&lib_path);
-
-        assert_eq!(app.library.as_ref().unwrap().snippets.len(), 2);
-        assert_eq!(app.library.as_ref().unwrap().snippets[0].title, "A");
-        assert_eq!(app.library.as_ref().unwrap().snippets[1].title, "C");
-        assert_eq!(app.library_selected, 1, "Selected index stays at 1 (now C)");
-
-        // Verify persisted
-        let lib = crate::library::load_library(&lib_path).unwrap();
-        assert_eq!(lib.snippets.len(), 2);
-    }
-
-    #[test]
-    fn d_on_last_item_adjusts_selection() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["A", "B"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-
-        // Select last item and delete
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.library_selected, 1);
-
-        app.delete_library_snippet_from(&lib_path);
-
-        assert_eq!(app.library.as_ref().unwrap().snippets.len(), 1);
-        assert_eq!(app.library_selected, 0, "Adjusted to last valid index");
-    }
-
-    #[test]
-    fn d_on_empty_library_is_noop() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-
-        assert!(app.library.as_ref().unwrap().snippets.is_empty());
-
-        app.delete_library_snippet_from(&lib_path);
-
-        assert!(app.library.as_ref().unwrap().snippets.is_empty());
-    }
-
-    #[test]
-    fn library_browse_loads_from_disk() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["X", "Y"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-
-        let lib = app.library.as_ref().unwrap();
-        assert_eq!(lib.snippets.len(), 2);
-        assert_eq!(lib.snippets[0].title, "X");
-        assert_eq!(lib.snippets[1].title, "Y");
-    }
-
-    // --- Rename tests ---
-
-    #[test]
-    fn r_in_library_browse_enters_rename_with_current_title() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["My Snippet"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-
-        app.handle_key_event(key_event(KeyCode::Char('r')));
-
-        assert_eq!(app.mode, Mode::RenameInput);
-        assert_eq!(app.title_input, "My Snippet");
-    }
-
-    #[test]
-    fn rename_esc_returns_to_library_browse() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::RenameInput;
-        app.title_input = "partial edit".to_string();
-
-        app.handle_key_event(key_event(KeyCode::Esc));
-
-        assert_eq!(app.mode, Mode::LibraryBrowse);
-        assert!(app.title_input.is_empty());
-    }
-
-    #[test]
-    fn rename_enter_saves_new_title() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["Old Title"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-        app.mode = Mode::RenameInput;
-        app.title_input = "New Title".to_string();
-
-        app.rename_library_snippet_from(&lib_path);
-
-        assert_eq!(app.mode, Mode::LibraryBrowse);
-        assert!(app.title_input.is_empty());
-        assert_eq!(app.library.as_ref().unwrap().snippets[0].title, "New Title");
-
-        // Verify persisted
-        let lib = crate::library::load_library(&lib_path).unwrap();
-        assert_eq!(lib.snippets[0].title, "New Title");
-    }
-
-    #[test]
-    fn rename_with_empty_title_shows_error() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-        library_with_snippets(&lib_path, &["Keep Me"]);
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-        app.mode = Mode::RenameInput;
-        app.title_input = "  ".to_string();
-
-        app.rename_library_snippet_from(&lib_path);
-
-        assert_eq!(app.mode, Mode::RenameInput, "Stays in RenameInput on empty");
-        assert!(app.status_message.as_deref().unwrap().contains("empty"));
-
-        // Original title preserved
-        let lib = crate::library::load_library(&lib_path).unwrap();
-        assert_eq!(lib.snippets[0].title, "Keep Me");
-    }
-
-    #[test]
-    fn r_on_empty_library_is_noop() {
-        let tmp = TempDir::new().unwrap();
-        let lib_path = tmp.path().join("library.toml");
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_library_browse_from(&lib_path);
-
-        app.handle_key_event(key_event(KeyCode::Char('r')));
-
-        assert_eq!(
-            app.mode,
-            Mode::LibraryBrowse,
-            "Stays in browse on empty lib"
-        );
-    }
-
-    // --- Settings screen tests ---
-
-    #[test]
-    fn app_starts_on_files_screen() {
-        let app = App::new(vec![], &Config::default());
-        assert_eq!(app.screen, Screen::Files);
-    }
-
-    #[test]
-    fn pressing_2_switches_to_settings() {
-        let mut app = App::new(vec![], &Config::default());
-        let collection = crate::settings::SettingsCollection {
-            files: vec![crate::settings::SettingsFile {
-                label: "Test".to_string(),
-                path: PathBuf::from("/test/settings.json"),
-                value: serde_json::json!({"model": "opus"}),
-            }],
-        };
-        app.switch_to_settings_with(&collection);
-        assert_eq!(app.screen, Screen::Settings);
-        assert!(!app.settings_state.lines.is_empty());
-    }
-
-    #[test]
-    fn pressing_1_returns_to_files() {
-        let mut app = App::new(vec![], &Config::default());
-        app.screen = Screen::Settings;
-        app.handle_key_event(key_event(KeyCode::Char('1')));
-        assert_eq!(app.screen, Screen::Files);
-    }
-
-    #[test]
-    fn pressing_2_in_title_input_types_char_not_switch() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::TitleInput;
-        app.handle_key_event(key_event(KeyCode::Char('2')));
-        assert_eq!(app.screen, Screen::Files, "Should NOT switch screen");
-        assert_eq!(app.title_input, "2", "Should type '2' into input");
-    }
-
-    #[test]
-    fn q_on_settings_exits() {
-        let mut app = App::new(vec![], &Config::default());
-        app.screen = Screen::Settings;
-        app.handle_key_event(key_event(KeyCode::Char('q')));
-        assert!(app.exit);
-    }
-
-    #[test]
-    fn jk_on_settings_scrolls() {
-        let mut app = App::new(vec![], &Config::default());
-        app.screen = Screen::Settings;
-        app.settings_state.lines = vec![
-            "Line 0".to_string(),
-            "Line 1".to_string(),
-            "Line 2".to_string(),
-            "Line 3".to_string(),
-        ];
-        app.settings_state.viewport_height = 10;
-
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.settings_state.cursor, 1);
-
-        app.handle_key_event(key_event(KeyCode::Char('j')));
-        assert_eq!(app.settings_state.cursor, 2);
-
-        app.handle_key_event(key_event(KeyCode::Char('k')));
-        assert_eq!(app.settings_state.cursor, 1);
-    }
-
-    // --- Edit mode tests ---
-
-    #[test]
-    fn edit_state_starts_as_none() {
-        let app = App::new(vec![], &Config::default());
-        assert!(app.edit_state.is_none());
-    }
-
-    #[test]
-    fn ctrl_c_exits_from_edit_mode() {
-        let mut app = App::new(vec![], &Config::default());
-        app.mode = Mode::Edit;
-        app.handle_key_event(KeyEvent {
-            code: KeyCode::Char('c'),
-            modifiers: KeyModifiers::CONTROL,
-            kind: KeyEventKind::Press,
-            state: KeyEventState::empty(),
-        });
-        assert!(app.exit, "Ctrl-C should exit from edit mode");
-    }
-
-    #[test]
-    fn edit_state_is_dirty_detects_changes() {
-        let original = "line 1\nline 2".to_string();
-        let textarea = TextArea::from(["line 1", "line 2"]);
-        let state = EditState {
-            textarea,
-            file_path: PathBuf::from("/test/file.md"),
-            original_text: original,
-            had_trailing_newline: false,
-            discard_confirmed: false,
-            dirty_cache: Cell::new(None),
-        };
-        assert!(!state.is_dirty(), "Unmodified textarea should not be dirty");
-
-        let modified_textarea = TextArea::from(["line 1", "line 2 modified"]);
-        let state2 = EditState {
-            textarea: modified_textarea,
-            file_path: PathBuf::from("/test/file.md"),
-            original_text: "line 1\nline 2".to_string(),
-            had_trailing_newline: false,
-            discard_confirmed: false,
-            dirty_cache: Cell::new(None),
-        };
-        assert!(state2.is_dirty(), "Modified textarea should be dirty");
-    }
-
-    #[test]
-    fn e_in_content_pane_enters_edit_mode() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Hello world").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file],
-        }];
-        let mut app = App::new(roots, &Config::default());
-        app.active_pane = Pane::Content;
-
-        app.handle_key_event(key_event(KeyCode::Char('e')));
-
-        assert_eq!(app.mode, Mode::Edit);
-        assert!(app.edit_state.is_some());
-        let edit = app.edit_state.as_ref().unwrap();
-        assert_eq!(edit.textarea.lines().join("\n"), "Hello world");
-        assert!(!edit.is_dirty());
-    }
-
-    #[test]
-    fn edit_mode_renders_without_panic() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Line 1\nLine 2\nLine 3").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file],
-        }];
-        let mut app = App::new(roots, &Config::default());
-        app.active_pane = Pane::Content;
-        app.handle_key_event(key_event(KeyCode::Char('e')));
-        assert_eq!(app.mode, Mode::Edit);
-
-        let backend = TestBackend::new(80, 24);
-        let mut terminal = Terminal::new(backend).unwrap();
-        terminal.draw(|frame| app.draw(frame)).unwrap();
-        // If we get here without panic, the test passes
-    }
-
-    #[test]
-    fn e_in_file_list_does_not_enter_edit() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Hello").unwrap();
-
-        let roots = vec![SourceRoot {
-            path: tmp.path().to_path_buf(),
-            files: vec![file],
-        }];
-        let mut app = App::new(roots, &Config::default());
-        app.active_pane = Pane::FileList;
-
-        app.handle_key_event(key_event(KeyCode::Char('e')));
-
-        assert_eq!(app.mode, Mode::Normal);
-        assert!(app.edit_state.is_none());
-    }
-
-    #[test]
-    fn typing_in_edit_mode_modifies_textarea() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Hello").unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-        assert_eq!(app.mode, Mode::Edit);
-
-        // Type a character
-        app.handle_key_event(key_event(KeyCode::Char('X')));
-
-        let edit = app.edit_state.as_ref().unwrap();
-        let content = edit.textarea.lines().join("\n");
-        assert!(content.contains('X'), "Typed char should appear: {content}");
-        assert!(edit.is_dirty());
-    }
-
-    #[test]
-    fn ctrl_s_saves_file_to_disk() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "original").unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-
-        // Type something
-        app.handle_key_event(key_event(KeyCode::Char('!')));
-
-        // Save with Ctrl+S
-        app.handle_key_event(KeyEvent {
-            code: KeyCode::Char('s'),
-            modifiers: KeyModifiers::CONTROL,
-            kind: KeyEventKind::Press,
-            state: KeyEventState::empty(),
-        });
-
-        let saved = fs::read_to_string(&file).unwrap();
-        assert!(
-            saved.contains('!'),
-            "File should contain typed char: {saved}"
-        );
-        assert!(app.status_message.as_deref().unwrap().contains("Saved"));
-
-        // Should still be in edit mode after save
-        assert_eq!(app.mode, Mode::Edit);
-        // But no longer dirty
-        assert!(
-            !app.edit_state.as_ref().unwrap().is_dirty(),
-            "After save, should not be dirty"
-        );
-    }
-
-    #[test]
-    fn esc_on_clean_edit_returns_to_normal() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "clean").unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-
-        // Esc with no changes should exit edit mode
-        app.handle_key_event(key_event(KeyCode::Esc));
-
-        assert_eq!(app.mode, Mode::Normal);
-        assert!(app.edit_state.is_none());
-    }
-
-    #[test]
-    fn esc_on_dirty_edit_warns_does_not_exit() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "original").unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-        app.handle_key_event(key_event(KeyCode::Char('X')));
-
-        // First Esc: warns but doesn't exit
-        app.handle_key_event(key_event(KeyCode::Esc));
-
-        assert_eq!(app.mode, Mode::Edit, "Should still be in edit mode");
-        assert!(
-            app.status_message.as_deref().unwrap().contains("unsaved"),
-            "Should show unsaved warning"
-        );
-    }
-
-    #[test]
-    fn double_esc_discards_unsaved_changes() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "original").unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-        app.handle_key_event(key_event(KeyCode::Char('X')));
-
-        // First Esc: warns
-        app.handle_key_event(key_event(KeyCode::Esc));
-        assert_eq!(app.mode, Mode::Edit);
-
-        // Second Esc: discards
-        app.handle_key_event(key_event(KeyCode::Esc));
-        assert_eq!(app.mode, Mode::Normal);
-        assert!(app.edit_state.is_none());
-
-        // File should be unchanged
-        let content = fs::read_to_string(&file).unwrap();
-        assert_eq!(content, "original");
-    }
-
-    #[test]
-    fn typing_after_first_esc_resets_discard_flag() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "original").unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-        app.handle_key_event(key_event(KeyCode::Char('X')));
-
-        // First Esc: warns
-        app.handle_key_event(key_event(KeyCode::Esc));
-        assert!(app.edit_state.as_ref().unwrap().discard_confirmed);
-
-        // Type something: resets discard flag
-        app.handle_key_event(key_event(KeyCode::Char('Y')));
-        assert!(!app.edit_state.as_ref().unwrap().discard_confirmed);
-
-        // Now Esc again should warn, not discard
-        app.handle_key_event(key_event(KeyCode::Esc));
-        assert_eq!(app.mode, Mode::Edit, "Should still be in edit mode");
-    }
-
     #[test]
     fn help_line_shows_edit_key_in_content_pane() {
         let mut app = App::new(vec![], &Config::default());
@@ -3109,18 +916,6 @@ mod tests {
     }
 
     #[test]
-    fn help_line_shows_edit_key_on_settings_screen() {
-        let mut app = App::new(vec![], &Config::default());
-        app.screen = Screen::Settings;
-        let help = app.help_line();
-        let help_text: String = help.spans.iter().map(|s| s.content.to_string()).collect();
-        assert!(
-            help_text.contains("Edit"),
-            "Help line should show Edit on settings screen: {help_text}"
-        );
-    }
-
-    #[test]
     fn help_line_shows_save_cancel_in_edit_mode() {
         let mut app = App::new(vec![], &Config::default());
         app.mode = Mode::Edit;
@@ -3129,368 +924,6 @@ mod tests {
         assert!(
             help_text.contains("Save") && help_text.contains("Cancel"),
             "Help line should show Save and Cancel in edit mode: {help_text}"
-        );
-    }
-
-    #[test]
-    fn e_on_settings_screen_enters_edit_for_settings_file() {
-        let tmp = TempDir::new().unwrap();
-        let settings_dir = tmp.path().join(".claude");
-        fs::create_dir_all(&settings_dir).unwrap();
-        let settings_file = settings_dir.join("settings.json");
-        fs::write(&settings_file, r#"{"model":"opus"}"#).unwrap();
-
-        let collection = crate::settings::SettingsCollection {
-            files: vec![crate::settings::SettingsFile {
-                label: "Test".to_string(),
-                path: settings_file.clone(),
-                value: serde_json::json!({"model": "opus"}),
-            }],
-        };
-
-        let mut app = App::new(vec![], &Config::default());
-        app.switch_to_settings_with(&collection);
-        app.settings_state.cursor = 0; // on the header line of the file
-
-        app.handle_key_event(key_event(KeyCode::Char('e')));
-
-        assert_eq!(app.mode, Mode::Edit);
-        assert!(app.edit_state.is_some());
-        let edit = app.edit_state.as_ref().unwrap();
-        assert_eq!(edit.file_path, settings_file);
-    }
-
-    #[test]
-    fn exiting_settings_edit_refreshes_formatted_view() {
-        let tmp = TempDir::new().unwrap();
-        let settings_dir = tmp.path().join(".claude");
-        fs::create_dir_all(&settings_dir).unwrap();
-        let settings_file = settings_dir.join("settings.json");
-        fs::write(&settings_file, r#"{"model":"opus"}"#).unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&settings_file);
-        app.screen = Screen::Settings;
-        assert_eq!(app.mode, Mode::Edit);
-
-        // Exit without changes
-        app.handle_key_event(key_event(KeyCode::Esc));
-        assert_eq!(app.mode, Mode::Normal);
-        assert!(app.edit_state.is_none());
-    }
-
-    #[test]
-    fn settings_file_at_cursor_resolves_path() {
-        let mut app = App::new(vec![], &Config::default());
-        let collection = crate::settings::SettingsCollection {
-            files: vec![
-                crate::settings::SettingsFile {
-                    label: "Global".to_string(),
-                    path: PathBuf::from("/home/.claude/settings.json"),
-                    value: serde_json::json!({"model": "opus"}),
-                },
-                crate::settings::SettingsFile {
-                    label: "Project".to_string(),
-                    path: PathBuf::from("/proj/.claude/settings.json"),
-                    value: serde_json::json!({"defaultMode": "plan"}),
-                },
-            ],
-        };
-        app.switch_to_settings_with(&collection);
-
-        // Cursor at 0 should be the Global file header
-        app.settings_state.cursor = 0;
-        assert_eq!(
-            app.settings_file_at_cursor(),
-            Some(Path::new("/home/.claude/settings.json"))
-        );
-
-        // Find a line from the second file
-        let second_file_line = app
-            .settings_state
-            .line_map
-            .iter()
-            .position(|m| *m == Some(1))
-            .unwrap();
-        app.settings_state.cursor = second_file_line;
-        assert_eq!(
-            app.settings_file_at_cursor(),
-            Some(Path::new("/proj/.claude/settings.json"))
-        );
-    }
-
-    #[test]
-    fn q_in_edit_mode_types_q_not_exit() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        fs::write(&file, "Hello").unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-
-        app.handle_key_event(key_event(KeyCode::Char('q')));
-
-        assert!(!app.exit, "q should not exit in edit mode");
-        assert_eq!(app.mode, Mode::Edit);
-        let edit = app.edit_state.as_ref().unwrap();
-        let content = edit.textarea.lines().join("\n");
-        assert!(content.contains('q'), "q should be typed into editor");
-    }
-
-    #[test]
-    fn enter_edit_mode_for_nonexistent_file_stays_normal() {
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(Path::new("/nonexistent/CLAUDE.md"));
-
-        assert_eq!(app.mode, Mode::Normal, "Should stay in Normal mode");
-        assert!(app.edit_state.is_none(), "No edit state should be created");
-        assert!(
-            app.status_message
-                .as_deref()
-                .unwrap()
-                .contains("Cannot open"),
-            "Should show error message, got: {:?}",
-            app.status_message
-        );
-    }
-
-    #[test]
-    fn e_on_blank_separator_in_settings_shows_error() {
-        let mut app = App::new(vec![], &Config::default());
-        let collection = crate::settings::SettingsCollection {
-            files: vec![
-                crate::settings::SettingsFile {
-                    label: "Global".to_string(),
-                    path: PathBuf::from("/home/.claude/settings.json"),
-                    value: serde_json::json!({"model": "opus"}),
-                },
-                crate::settings::SettingsFile {
-                    label: "Project".to_string(),
-                    path: PathBuf::from("/proj/.claude/settings.json"),
-                    value: serde_json::json!({"defaultMode": "plan"}),
-                },
-            ],
-        };
-        app.switch_to_settings_with(&collection);
-
-        // Find the blank separator line (maps to None)
-        let blank_idx = app
-            .settings_state
-            .line_map
-            .iter()
-            .position(|m| m.is_none())
-            .unwrap();
-        app.settings_state.cursor = blank_idx;
-
-        app.handle_key_event(key_event(KeyCode::Char('e')));
-
-        assert_eq!(app.mode, Mode::Normal, "Should stay in Normal mode");
-        assert!(app.edit_state.is_none());
-        assert!(
-            app.status_message
-                .as_deref()
-                .unwrap()
-                .contains("No settings file"),
-            "Should show no-file message, got: {:?}",
-            app.status_message
-        );
-    }
-
-    fn two_file_settings_collection() -> crate::settings::SettingsCollection {
-        crate::settings::SettingsCollection {
-            files: vec![
-                crate::settings::SettingsFile {
-                    label: "Global".to_string(),
-                    path: PathBuf::from("/home/.claude/settings.json"),
-                    value: serde_json::json!({"model": "opus", "permissions": {"allow": ["Read"]}}),
-                },
-                crate::settings::SettingsFile {
-                    label: "Project".to_string(),
-                    path: PathBuf::from("/proj/.claude/settings.json"),
-                    value: serde_json::json!({"model": "haiku", "permissions": {"allow": ["Write"]}}),
-                },
-            ],
-        }
-    }
-
-    #[test]
-    fn m_key_toggles_merged_view() {
-        let mut app = App::new(vec![], &Config::default());
-        app.switch_to_settings_with(&two_file_settings_collection());
-
-        // Per-file view has two section headers
-        let headers_before: Vec<_> = app
-            .settings_state
-            .lines
-            .iter()
-            .filter(|l| l.starts_with('\u{25be}'))
-            .collect();
-        assert_eq!(
-            headers_before.len(),
-            2,
-            "Per-file view should have 2 headers"
-        );
-
-        // Toggle to merged
-        app.handle_key_event(key_event(KeyCode::Char('m')));
-        assert!(app.settings_state.merged_view);
-
-        let headers_after: Vec<_> = app
-            .settings_state
-            .lines
-            .iter()
-            .filter(|l| l.starts_with('\u{25be}'))
-            .collect();
-        assert_eq!(headers_after.len(), 1, "Merged view should have 1 header");
-        assert!(
-            headers_after[0].contains("Effective"),
-            "Header should say Effective, got: {}",
-            headers_after[0]
-        );
-    }
-
-    #[test]
-    fn m_key_resets_cursor() {
-        let mut app = App::new(vec![], &Config::default());
-        app.switch_to_settings_with(&two_file_settings_collection());
-
-        app.settings_state.cursor = 5;
-        app.settings_state.scroll = 3;
-
-        app.handle_key_event(key_event(KeyCode::Char('m')));
-
-        assert_eq!(
-            app.settings_state.cursor, 0,
-            "Cursor should reset on toggle"
-        );
-        assert_eq!(
-            app.settings_state.scroll, 0,
-            "Scroll should reset on toggle"
-        );
-    }
-
-    #[test]
-    fn m_key_round_trip() {
-        let mut app = App::new(vec![], &Config::default());
-        app.switch_to_settings_with(&two_file_settings_collection());
-        let lines_before = app.settings_state.lines.clone();
-
-        // Toggle to merged, then back
-        app.handle_key_event(key_event(KeyCode::Char('m')));
-        app.handle_key_event(key_event(KeyCode::Char('m')));
-
-        assert!(!app.settings_state.merged_view);
-        assert_eq!(app.settings_state.lines, lines_before);
-    }
-
-    #[test]
-    fn e_disabled_in_merged_view() {
-        let mut app = App::new(vec![], &Config::default());
-        app.switch_to_settings_with(&two_file_settings_collection());
-        app.screen = Screen::Settings;
-
-        // Toggle to merged
-        app.handle_key_event(key_event(KeyCode::Char('m')));
-
-        // Press e
-        app.handle_key_event(key_event(KeyCode::Char('e')));
-
-        assert_eq!(
-            app.mode,
-            Mode::Normal,
-            "e should not enter edit in merged view"
-        );
-        assert!(
-            app.status_message
-                .as_deref()
-                .unwrap_or("")
-                .contains("merged view"),
-            "Should show merged view message, got: {:?}",
-            app.status_message
-        );
-    }
-
-    #[test]
-    fn help_bar_shows_merge_key() {
-        let mut app = App::new(vec![], &Config::default());
-        app.screen = Screen::Settings;
-        let help = app.help_line();
-        let help_text: String = help.spans.iter().map(|s| s.content.to_string()).collect();
-        assert!(
-            help_text.contains("Merge"),
-            "Help bar should show Merge key in per-file view: {help_text}"
-        );
-        assert!(
-            help_text.contains("Edit"),
-            "Help bar should show Edit in per-file view: {help_text}"
-        );
-    }
-
-    #[test]
-    fn help_bar_in_merged_omits_edit() {
-        let mut app = App::new(vec![], &Config::default());
-        app.screen = Screen::Settings;
-        app.settings_state.merged_view = true;
-        let help = app.help_line();
-        let help_text: String = help.spans.iter().map(|s| s.content.to_string()).collect();
-        assert!(
-            help_text.contains("Per-file"),
-            "Help bar should show Per-file in merged view: {help_text}"
-        );
-        assert!(
-            !help_text.contains("Edit"),
-            "Help bar should NOT show Edit in merged view: {help_text}"
-        );
-    }
-
-    #[test]
-    fn trailing_newline_preserved_after_edit_save_cycle() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        let original_content = "Line 1\nLine 2\n";
-        fs::write(&file, original_content).unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-        assert_eq!(app.mode, Mode::Edit);
-
-        // Save without making changes
-        app.handle_key_event(KeyEvent {
-            code: KeyCode::Char('s'),
-            modifiers: KeyModifiers::CONTROL,
-            kind: KeyEventKind::Press,
-            state: KeyEventState::empty(),
-        });
-
-        let saved = fs::read_to_string(&file).unwrap();
-        assert_eq!(
-            saved, original_content,
-            "File should be byte-for-byte identical after no-op save"
-        );
-    }
-
-    #[test]
-    fn no_trailing_newline_preserved_after_edit_save_cycle() {
-        let tmp = TempDir::new().unwrap();
-        let file = tmp.path().join("CLAUDE.md");
-        let original_content = "Line 1\nLine 2";
-        fs::write(&file, original_content).unwrap();
-
-        let mut app = App::new(vec![], &Config::default());
-        app.enter_edit_mode_for(&file);
-
-        // Save without making changes
-        app.handle_key_event(KeyEvent {
-            code: KeyCode::Char('s'),
-            modifiers: KeyModifiers::CONTROL,
-            kind: KeyEventKind::Press,
-            state: KeyEventState::empty(),
-        });
-
-        let saved = fs::read_to_string(&file).unwrap();
-        assert_eq!(
-            saved, original_content,
-            "File without trailing newline should stay without one"
         );
     }
 
@@ -3559,281 +992,6 @@ mod tests {
         };
         app.handle_key_event(shift_t);
         assert!(!app.theme.is_dark, "Theme should toggle on settings screen");
-    }
-
-    fn settings_app_with_lines(lines: Vec<&str>) -> App {
-        let mut app = App::new(vec![], &Config::default());
-        app.screen = Screen::Settings;
-        app.settings_state.lines = lines.into_iter().map(String::from).collect();
-        app.settings_state.line_map = vec![Some(0); app.settings_state.lines.len()];
-        app
-    }
-
-    // --- Fold/unfold tests ---
-
-    #[test]
-    fn is_foldable_detects_lines_with_children() {
-        let app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  Model: opus",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx rust-cargo",
-        ]);
-        assert!(app.settings_state.is_foldable(0), "Top header is foldable");
-        assert!(!app.settings_state.is_foldable(1), "Leaf is not foldable");
-        assert!(app.settings_state.is_foldable(2), "Sub-header is foldable");
-        assert!(!app.settings_state.is_foldable(3), "Leaf is not foldable");
-    }
-
-    #[test]
-    fn parent_for_returns_nearest_ancestor() {
-        let app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        assert_eq!(app.settings_state.parent_for(0), None);
-        assert_eq!(app.settings_state.parent_for(1), Some(0));
-        assert_eq!(app.settings_state.parent_for(2), Some(1));
-    }
-
-    #[test]
-    fn fold_top_level_hides_all_children() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  Model: opus",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        app.settings_state.toggle_fold(0);
-        assert!(
-            app.settings_state.is_line_visible(0),
-            "Header stays visible"
-        );
-        assert!(!app.settings_state.is_line_visible(1));
-        assert!(!app.settings_state.is_line_visible(2));
-        assert!(!app.settings_state.is_line_visible(3));
-
-        app.settings_state.toggle_fold(0);
-        assert!(app.settings_state.is_line_visible(3), "All visible again");
-    }
-
-    #[test]
-    fn fold_sub_section_hides_only_its_children() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  Model: opus",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-            "    github: gh",
-            "  Thinking: true",
-        ]);
-        // Fold "  ▾ MCP Servers:"
-        app.settings_state.toggle_fold(2);
-        assert!(app.settings_state.is_line_visible(0));
-        assert!(app.settings_state.is_line_visible(1), "Model still visible");
-        assert!(app.settings_state.is_line_visible(2), "MCP header visible");
-        assert!(!app.settings_state.is_line_visible(3), "rust-cargo hidden");
-        assert!(!app.settings_state.is_line_visible(4), "github hidden");
-        assert!(
-            app.settings_state.is_line_visible(5),
-            "Thinking still visible"
-        );
-    }
-
-    #[test]
-    fn cursor_down_skips_collapsed_lines() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  Model: opus",
-            "  Thinking: true",
-            "▾ Project (/other)",
-            "  Model: sonnet",
-        ]);
-        app.settings_state.toggle_fold(0);
-        app.settings_state.cursor = 0;
-
-        app.settings_state.cursor_down();
-        assert_eq!(app.settings_state.cursor, 3, "Should skip to next header");
-    }
-
-    #[test]
-    fn cursor_up_skips_collapsed_lines() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  Model: opus",
-            "  Thinking: true",
-            "▾ Project (/other)",
-            "  Model: sonnet",
-        ]);
-        app.settings_state.toggle_fold(0);
-        app.settings_state.cursor = 3;
-
-        app.settings_state.cursor_up();
-        assert_eq!(app.settings_state.cursor, 0, "Should skip back to header");
-    }
-
-    #[test]
-    fn cursor_skips_sub_section_collapsed_lines() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-            "    github: gh",
-            "  Thinking: true",
-        ]);
-        app.settings_state.toggle_fold(1); // Fold MCP Servers
-        app.settings_state.cursor = 1;
-
-        app.settings_state.cursor_down();
-        assert_eq!(app.settings_state.cursor, 4, "Should skip to Thinking");
-    }
-
-    #[test]
-    fn left_arrow_on_foldable_collapses() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        app.settings_state.cursor = 1; // On "  ▾ MCP Servers:"
-
-        app.handle_key_event(key_event(KeyCode::Left));
-        assert!(
-            app.settings_state.collapsed.contains(&1),
-            "Sub-section should be collapsed"
-        );
-    }
-
-    #[test]
-    fn left_arrow_on_leaf_jumps_to_parent() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        app.settings_state.cursor = 2; // On "    rust-cargo: npx"
-
-        app.handle_key_event(key_event(KeyCode::Left));
-        assert_eq!(
-            app.settings_state.cursor, 1,
-            "Should jump to MCP Servers parent"
-        );
-    }
-
-    #[test]
-    fn left_on_collapsed_foldable_jumps_to_parent() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        app.settings_state.toggle_fold(1); // Already collapsed
-        app.settings_state.cursor = 1;
-
-        app.handle_key_event(key_event(KeyCode::Left));
-        assert_eq!(
-            app.settings_state.cursor, 0,
-            "Should jump to top-level parent"
-        );
-    }
-
-    #[test]
-    fn right_arrow_on_collapsed_sub_section_expands() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        app.settings_state.toggle_fold(1);
-        app.settings_state.cursor = 1;
-
-        app.handle_key_event(key_event(KeyCode::Right));
-        assert!(
-            !app.settings_state.collapsed.contains(&1),
-            "Sub-section should be expanded"
-        );
-    }
-
-    #[test]
-    fn right_arrow_on_expanded_is_noop() {
-        let mut app = settings_app_with_lines(vec!["▾ Global (/path)", "  Model: opus"]);
-        app.settings_state.cursor = 0;
-
-        app.handle_key_event(key_event(KeyCode::Right));
-        assert!(!app.settings_state.collapsed.contains(&0));
-    }
-
-    #[test]
-    fn visible_line_count_respects_collapsed() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  Model: opus",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-            "    github: gh",
-        ]);
-        assert_eq!(app.settings_state.visible_line_count(), 5);
-
-        app.settings_state.toggle_fold(2); // Fold MCP Servers
-        assert_eq!(app.settings_state.visible_line_count(), 3);
-
-        app.settings_state.toggle_fold(0); // Fold entire Global
-        assert_eq!(app.settings_state.visible_line_count(), 1);
-    }
-
-    #[test]
-    fn nested_fold_parent_hides_expanded_children() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        // Children of MCP Servers are expanded, but fold the parent
-        app.settings_state.toggle_fold(0);
-        assert!(!app.settings_state.is_line_visible(1), "MCP Servers hidden");
-        assert!(!app.settings_state.is_line_visible(2), "rust-cargo hidden");
-    }
-
-    #[test]
-    fn rebuild_settings_clears_collapsed_state() {
-        let mut app = settings_app_with_lines(vec!["▾ Global (/path)", "  Model: opus"]);
-        app.settings_state.collapsed.insert(0);
-
-        app.settings_state.collapsed.clear();
-        assert!(app.settings_state.collapsed.is_empty());
-    }
-
-    #[test]
-    fn sub_header_indicator_toggles_on_fold() {
-        let mut app = settings_app_with_lines(vec![
-            "▾ Global (/path)",
-            "  ▾ MCP Servers:",
-            "    rust-cargo: npx",
-        ]);
-        assert!(app.settings_state.lines[1].contains('▾'));
-
-        app.settings_state.toggle_fold(1);
-        assert!(
-            app.settings_state.lines[1].contains('▸'),
-            "Should show collapsed indicator"
-        );
-
-        app.settings_state.toggle_fold(1);
-        assert!(
-            app.settings_state.lines[1].contains('▾'),
-            "Should show expanded indicator again"
-        );
-    }
-
-    #[test]
-    fn top_header_indicator_toggles_on_fold() {
-        let mut app = settings_app_with_lines(vec!["▾ Global (/path)", "  Model: opus"]);
-        app.settings_state.toggle_fold(0);
-        assert!(app.settings_state.lines[0].starts_with('▸'));
-
-        app.settings_state.toggle_fold(0);
-        assert!(app.settings_state.lines[0].starts_with('▾'));
     }
 
     #[test]
