+++ src/tui/mod.rs
@@ -1,2 +1,5 @@
 pub mod app;
+pub mod edit;
+pub mod files;
+pub mod settings;
 pub mod theme;
